backend/requirements.txt
requests==2.32.4
jsonschema==4.24.0
liter-llm==1.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
numpy>=1.24.0
rank-bm25>=0.2.2
jieba>=0.42.1
